#401
- box: добавлена точка входа
- text: добавлена точка входа
- touch-slider: добавлена точка входа
