+++ src/box/index.ts
@@ -1,0 +1,2 @@
+export type { BoxProps } from './types';
+export { Box } from './box';
