+++ src/text/index.ts
@@ -1,0 +1,2 @@
+export type { TextProps } from './types';
+export { Text, defaultHeights, ALIGNS } from './text';
