+++ src/touch-slider/index.ts
@@ -1,0 +1,2 @@
+export type { TouchSliderProps } from './types';
+export { TouchSlider } from './touch-slider';
