+++ src/touch-slider/types.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from 'react';
+
+export interface TouchSliderProps {
+  /** Прокручиваемое содержимое. */
+  children?: ReactNode;
+}
